Move Sourc, Sink to jms-client module
