+++ jms-client/src/main/resources/sink.properties
