+++ jms-client/src/main/resources/source.properties
