+++ routing-configurer/src/test/java/no/vegvesen/ixn/federation/qpid/QpidClientIT.java
@@ -1,7 +1,7 @@
 package no.vegvesen.ixn.federation.qpid;
 
-import no.vegvesen.interchange.Sink;
-import no.vegvesen.interchange.Source;
+import no.vegvesen.ixn.Sink;
+import no.vegvesen.ixn.Source;
 import no.vegvesen.ixn.federation.model.*;
 import no.vegvesen.ixn.ssl.KeystoreDetails;
 import no.vegvesen.ixn.ssl.KeystoreType;
